debugging

+++ src/main/java/miniplc0java/analyser/Analyser.java
@@ -719,7 +719,7 @@ public final class Analyser {
             }
         } else if (check(TokenType.UINT_LITERAL)) {
             Token name=expect(TokenType.UINT_LITERAL);
-            cur_func.getInstructions().add(new Instruction(Operation.PUSH,((Integer)name.getValue()).longValue()));
+            cur_func.getInstructions().add(new Instruction(Operation.PUSH,(Long)name.getValue()));
             type=IdentType.INT;
 
         } else if (check(TokenType.DOUBLE_LITERAL)) {
